removes unused variables, removes unused files, removes whitespace, restructures with modules
