+++ api/USERSTORY.md
@@ -1,3 +1,20 @@
 As a user, I'd like to see the player's stats from their last five games, so that I can view their most recent metrics
 
 As a user, I want to have a player feed that tracks real-time tickers, so that I can identify the plays they make frequently
+
+* TODO:
+TASKS:
+  * Style homepage
+  * Add conference standings to home page
+
+FEATURES:
+  * Player Transaction Feed on Home Page
+  
+  * Add Comparison Analytics
+    * Player Career Stats
+      * if trend is +, display the text as green
+      * if trend is -, display the text as red
+    * Render All Seasons to Table
+  * Add Shotchart
+
+  * Add conference standings to home page
