+++ api/views/index.ejs
@@ -16,6 +16,5 @@
   <%- include('../views/feed.ejs') %>
 </div>
 <% }; %>
-  <%- include('../views/footer.ejs') %>
 </body>
 </html>
